서버에서 클라이언트를 기다리는 메시지 출력 추가 및 파이프 열기 순서 수정

+++ network/pipe_server.c
@@ -13,20 +13,32 @@ int main()
 {
     int fd1, fd2;
     char buffer[BUFFER_SIZE];
-
     mkfifo(PIPE1, 0666);
     mkfifo(PIPE2, 0666);
 
-    fd1 = open(PIPE1, O_WRONLY);
-    fd2 = open(PIPE2, O_RDONLY);
+    printf("클라이언트 기다림");
 
-    char *message = "Hello from Parent!";
+    fd1 = open(PIPE1, O_WRONLY);
+    if (fd1 == -1)
+    {
+        perror("open pipe1");
+        exit(EXIT_FAILURE);
+    }
+    printf("테스트3:\n");
+    char *message = "Hello from Server!";
     write(fd1, message, strlen(message) + 1);
     printf("서버에서 보낸 메세지: %s\n", message);
-    getchar();
+    close(fd1);
+
+    fd2 = open(PIPE2, O_RDONLY);
+    if (fd2 == -1)
+    {
+        perror("open pipe2");
+        exit(EXIT_FAILURE);
+    }
+    printf("테스트4:\n");
     read(fd2, buffer, BUFFER_SIZE); // 대기
     printf("서버에서 읽은 메세지: %s\n", buffer);
-    close(fd1);
     close(fd2);
 
     unlink(PIPE1);
